Alter function so user can ONLY enter R, P or S
Before this, there was no way for the function to evaluate an input from
the user if the input was NOT Rock, Paper or Scissors. This function has
now been corrected so that if the input is neither Rock, Paper nor
Scissors it returns a statement advising the user to enter either Rock,
Paper or Scissors.