Alter singleRoundOfRPS parameter to include prompt
Before this, the singleRoundOfRPS function did not contain the prompt
as a parameter. This produced errors when it came to trying to get the
game function to work as this required a prompt box with a different
input, each, for 5 different games instead of just one. The game
function calls upon the singleRoundOfRPS function.