Fix function to run RPS 5 times and report result
Previously an incorrect attempt of using for loops was present. Instead
of using for loops, the function is mow called upon 5 separate times.
The result of each separate game adds 1 to the win/lose/draw variable.
If either the win or lose variable is > 2 it reports the outcome with a
corresponding win or lose statement otherwise it returns a draw
statement