Fix singleRoundOfRPS function if RPS not entered
Before, the user could enter anything other than RPS 5 times and this
would take up the game function. The result of this would be 'draw' even
though this was not the case. To correct this, in the singleRoundOfRPS
function, if a player does not enter RPS the function is returned and
thus repeated until a person enters R P or S.