Show outcome of each individual game in game func
Before this, the game would call the singleRoundOfRPS function 5 times
and then ultimately declare and return a winner/loser/draw statement.
This meant that the person playing could not see how many games they
had won and how and when they won/lost them during the 5 games. Now the
player can see the outcomes of each individual game in the 5 game set to
prepare for being victorious or defeat. This includes if a person enters
something other than RPS - they can now see it but also a
singleRoundOfRPS function is returned if they do. So that if they enter
something other than RPS it does not take up a function/game.